+++ src/components/ImagePicker.vue
